Async download of lyrics & maps!

+++ app/src/main/java/io/github/dkambersky/songle/data/definitions/SongleContext.kt
@@ -8,7 +8,7 @@ import java.io.Serializable
  * Carries application data for easy passing around
  */
 data class SongleContext(val songs: MutableList<Song>,
-                         val maps: MutableMap<Short, MutableMap<Short, List<Placemark>>>,
+                         val maps: MutableMap<Int, MutableMap<Int, List<Placemark>>>,
                          val styles: MutableMap<String, Style>,
                          @Transient val context: Context,
                          var clearedSongs: MutableSet<Song>,
